requirements.txt
uvicorn==0.17.6
sqlalchemy~=1.3.13
passlib~=1.7.4
bcrypt==3.2.0
pydantic~=1.8.2
python-jose
fastapi~=0.65.2
datetime~=4.3